Add tests to genetic::population and use .tpp's
I changed eveything so we can use our ::genetic library
as an header-only library. That includes using .tpp files
(only a naming convention as they are really just .cpp files)
and includign them in the .hpp files, that way the implementation
is always in the headers of templated classes and we don't get
undefined reference errors. It is an alternative to what I did
in the previous commit, only now, the library does not need to
guess with which types the library is going to be used.