Change constraint and objective handling
The idea is the following, objectives and constraints have to be
tightly coupled because if they are not, then transgressing solutions
will never be dominated because their objectives may be reaching
very low levels due to the constraints being violated. When the
constraints are not respected, the objective must suffer too.