Change one-of things to singletons
	* urban::grid
	* urban::metro_network
	* urban::walking_network
	* urban::odx_matrix
	* osm_net::osm_net
	* transit_problem::route_pool

Basically, things that would be created only once and the passed
aroung as arguments are all singletons now. This, of course,
implied revamping many function headers. In the end, everything
is simpler.
I also changed the items received in the constraints and objective
functions not to be const, because things like the bus_network
may require evaluations when the algorithm computes the objectives.