Solve bug with tranfer pernalty in path finding
This commit also allowed me to discover the bug regarding
addition and removal of routes.
The problem is that the path finding algorithm does not allow
the simultaneous exploration of two routes between the same
stops, which makes it so that the result of the path finding
algorithm is dependent of the exploration order of the different
edges between the same stops.